# examples/gameroom/daemon/openapi.yml
...
openapi: '3.0.0'
info:
  version: 0.5.1
  title: Gameroomd API
  description: REST API for Gameroomd
servers:
  - url: http://localhost:8000
paths:
...
  /message/{circuit_id}/messages:
    get:
      tags:
        - Messages
      description: List all messages within the specified circuit
      parameters:
        - name: circuit_id
          in: path
          description: Circuit id of the XO games to list
          required: true
          schema:
            type: string
      responses:
        200:
          description: list of XO games
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApiMessage'
                  paging:
                    $ref: '#/components/schemas/Paging'
        500:
          description: Internal server error occurred
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        503:
          description: Server unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

#  /xo/{circuit_id}/games/{game_name}:
#    get:
#      tags:
#        - XO Games
#      description: Fetch the XO game within the specified circuit and with the specified game name
#      parameters:
#        - name: circuit_id
#          in: path
#          description: Circuit id of the XO games to fetch
#          required: true
#          schema:
#            type: string
#        - name: game_name
#          in: path
#          description: Name of the XO game to fetch
#          required: true
#          schema:
#            type: string
#      responses:
#        200:
#          description: XO game
#          content:
#            application/json:
#              schema:
#                type: object
#                properties:
#                  data:
#                    $ref: '#/components/schemas/ApiXoGame'
#        404:
#          description: The XO game with {circuit_id} and {game_name} was not found
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/Error'
#        500:
#          description: Internal server error occurred
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/Error'
#        503:
#          description: Server unavailable
#          content:
#            application/json:
#              schema:
#                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      additionalProperties: false
      properties:
        code:
          description: The HTTP error code
          type: string
          example: 404
        message:
          description: A message describing the error that occurred
          type: string
          example: Resource not found
      required:
        - message
...
    Paging:
      type: object
      properties:
        current:
          type: string
        offset:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        first:
          type: string
        prev:
          type: string
        next:
          type: string
        last:
          type: string
      example:
        current: /nodes?offset=10&limit=10
        offset: 10
        limit: 10
        total: 50
        first: /nodes?offset=0&limit=10
        prev: /nodes?offset=0&limit=10
        next: /nodes?offset=20&limit=10
        last: /nodes?offset=40&limit=10
...
    ApiXoGame:
      type: object
      properties:
        circuit_id:
          type: string
        game_name:
          type: string
        player_1:
          type: string
        player_2:
          type: string
        game_status:
          type: string
        game_board:
          description: string representation of game board, using comma-separated values format
          type: string
        created_time:
          description: time in seconds since the unix epoch the XO game was created
          type: integer
        updated_time:
          description: time in seconds since the unix epoch the XO game was updated
          type: integer
...